+++ public/js/bird-vue-core.js
@@ -425,6 +425,7 @@
     weather:      'nav_weather',
     log:          'nav_log',
     purge:        'nav_purge',
+    quality:      'nav_quality',
     liveboard:    'nav_liveboard',
     dashboard_kiosk: 'nav_dashboard_kiosk',
     network:      'nav_network',